refactor: extract crates/worktree-core/ from git.rs (ADR-013 step 1)
Behavior-preserving refactor that pulls worktree mutation primitives out
of crates/orchard/src/git.rs into a standalone library crate. Backs the
forthcoming orchard-worktree CLI binary and (in step 4) the TUI dialogs,
giving the codebase a single source of truth for git worktree ops.

What moved (orchard → worktree-core):
- find_repo_root, get_repo_name → repo.rs
- list_worktrees, parse_porcelain, worktree_has_conflicts → list.rs
- remove_worktree → destroy.rs
- create_worktree (new, extracted from start_create_worktree) → create.rs

Type model: worktree-core uses its own minimal `WorktreeEntry` (just the
fields produced by `git worktree list --porcelain`). Orchard's enriched
`types::Worktree` (PR/tmux/issue joins) stays in the orchard crate where
it belongs.

Callsite updates:
- tui/mod.rs: git::find_repo_root → worktree_core::find_repo_root
- tui/worktree_ops.rs: git::remove_worktree → worktree_core::remove_worktree
- doc comment refs in git_parse.rs, logger.rs

ADR-013 marked ACCEPTED; docs/architecture.md adds the workspace-crates
section and the entry in the ADR list. Steps 2-6 (dispatcher binary,
orchard-worktree CLI, TUI migration, daemon rename, namespaced grammar)
follow in subsequent PRs.

Tests: 11 worktree-core unit tests pass. Pre-existing flakes in
sources::hosts and json_freshness_integration are unrelated (filed as #439).
