+++ crates/worktree-core/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "worktree-core"
+version = "0.1.0"
+edition = "2024"
+description = "Core git worktree operations: list, parse, create, remove, conflict detection"
+license = "MIT"
+repository = "https://github.com/drewdrewthis/git-orchard-rs"
+
+[lib]
+name = "worktree_core"
+path = "src/lib.rs"
+
+[dependencies]
+anyhow = "1"
+serde = { version = "1", features = ["derive"] }
